libft

+++ libft/src/ft_strncmp.c
@@ -1,7 +1,26 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_strncmp.c                                       :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: dthalman <daniel@thalmann.li>              +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2021/10/22 10:07:06 by dthalman          #+#    #+#             */
+/*   Updated: 2021/10/22 10:07:06 by dthalman         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strncmp(const char *s1, const char *s2, unsigned int n)
+/**
+ * @brief compares only the first (at most) n bytes of s1 and s2.
+ * 
+ * @param s1 
+ * @param s2 
+ * @param n 
+ * @return int 
+ */
+int	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
 	if (n == 0)
 		return (0);
